refactor: rename createApiClientEmbed to createApiClientApp
