+++ packages/api-client/package.json
@@ -25,8 +25,8 @@
   "scripts": {
     "build": "vite build && pnpm types:build && tsc-alias -p tsconfig.build.json",
     "dev": "vite",
-    "dev:playground:embed": "vite ./playground/embed -c ./vite.config.ts",
-    "dev:playground:modal": "vite ./playground/modal -c ./vite.config.ts",
+    "playground:app": "vite ./playground/app -c ./vite.config.ts",
+    "playground:modal": "vite ./playground/modal -c ./vite.config.ts",
     "lint:check": "eslint .",
     "lint:fix": "eslint .  --fix",
     "preview": "vite preview",
@@ -93,9 +93,9 @@
       "import": "./dist/Modal/index.js",
       "types": "./dist/Modal/index.d.ts"
     },
-    "./Embed": {
-      "import": "./dist/Embed/index.js",
-      "types": "./dist/Embed/index.d.ts"
+    "./App": {
+      "import": "./dist/App/index.js",
+      "types": "./dist/App/index.d.ts"
     },
     "./css/*.css": {
       "import": "./dist/css/*.css",
