+++ packages/api-client/playground/app/index.html
@@ -9,7 +9,7 @@
     <meta
       name="viewport"
       content="width=device-width, initial-scale=1.0" />
-    <title>Scalar API Client Modal</title>
+    <title>Scalar API Client App</title>
   </head>
   <body>
     <div
