+++ packages/api-client/src/App/createApiClientApp.ts
@@ -1,58 +1,16 @@
+import type { ClientConfiguration } from '@/Modal'
 import { appRouter } from '@/router'
 import { useWorkspace } from '@/store/workspace'
 import type { SpecConfiguration } from '@scalar/oas-utils'
 import { type RequestMethod, objectMerge } from '@scalar/oas-utils/helpers'
-import type { ThemeId } from '@scalar/themes'
 import { createApp } from 'vue'
 
-import ApiClientEmbed from './ApiClientEmbed.vue'
-
-/** Configuration options for the Scalar API client */
-export type ClientConfiguration = {
-  /** The Swagger/OpenAPI spec to render */
-  spec: SpecConfiguration
-  /** Pass in a proxy to the API client */
-  proxyUrl?: string
-  /** Pass in a theme API client */
-  themeId?: ThemeId
-  /** Whether to show the sidebar */
-  showSidebar?: boolean
-  /** Whether dark mode is on or off initially (light mode) */
-  // darkMode?: boolean
-  /** Key used with CTRL/CMD to open the search modal (defaults to 'k' e.g. CMD+k) */
-  searchHotKey?:
-    | 'a'
-    | 'b'
-    | 'c'
-    | 'd'
-    | 'e'
-    | 'f'
-    | 'g'
-    | 'h'
-    | 'i'
-    | 'j'
-    | 'k'
-    | 'l'
-    | 'm'
-    | 'n'
-    | 'o'
-    | 'p'
-    | 'q'
-    | 'r'
-    | 's'
-    | 't'
-    | 'u'
-    | 'v'
-    | 'w'
-    | 'x'
-    | 'y'
-    | 'z'
-}
+import ApiClientApp from './ApiClientApp.vue'
 
 export type OpenClientPayload = { path: string; method: RequestMethod }
 
 /** Initialize Scalar API Client embed */
-export const createApiClientEmbed = async (
+export const createApiClientApp = async (
   /** Element to mount the references to */
   el: HTMLElement | null,
   /** Configuration object for Scalar References */
@@ -67,7 +25,6 @@ export const createApiClientEmbed = async (
     activeWorkspace,
     importSpecFile,
     importSpecFromUrl,
-    requests,
     workspaceMutators,
   } = useWorkspace()
 
@@ -78,19 +35,19 @@ export const createApiClientEmbed = async (
     await importSpecFile(config.spec?.content)
   } else {
     console.error(
-      `[createApiClientEmbed] Could not create the API client.`,
+      `[createApiClientApp] Could not create the API client.`,
       `Please provide an OpenAPI document: { spec: { url: '…' } }`,
       `Read more: https://github.com/scalar/scalar/tree/main/packages/api-client-modal`,
     )
   }
 
-  const app = createApp(ApiClientEmbed)
+  const app = createApp(ApiClientApp)
   app.use(appRouter)
 
   const mount = (mountingEl = el) => {
     if (!mountingEl) {
       console.error(
-        `[createApiClientEmbed] Could not create the API client.`,
+        `[createApiClientApp] Could not create the API client.`,
         `Invalid HTML element provided.`,
         `Read more: https://github.com/scalar/scalar/tree/main/packages/api-client-modal`,
       )
@@ -140,23 +97,11 @@ export const createApiClientEmbed = async (
         importSpecFile(spec?.content)
       } else {
         console.error(
-          `[createApiClientEmbed] Could not create the API client.`,
+          `[createApiClientApp] Could not create the API client.`,
           `Please provide an OpenAPI document: { spec: { url: '…' } }`,
           `Read more: https://github.com/scalar/scalar/tree/main/packages/api-client-modal`,
         )
       }
-    },
-    /** Open the  API client modal */
-    open: (payload?: OpenClientPayload) => {
-      // Find the request from path + method
-      const request = Object.values(requests).find(({ path, method }) =>
-        payload
-          ? // The given operation
-            path === payload.path && method === payload.method
-          : // Or the first request
-            true,
-      )
-      if (request) appRouter.push(`/request/${request.uid}`)
     },
     /** Mount the references to a given element */
     mount,
