+++ packages/api-client/src/App/index.ts
@@ -1,0 +1,2 @@
+export { default as ApiClientApp } from './ApiClientApp.vue'
+export * from './createApiClientApp'
