--- packages/api-client/src/Embed/index.ts
@@ -1,2 +1,0 @@
-export { default as ApiClientEmbed } from './ApiClientEmbed.vue'
-export * from './createApiClientEmbed'
